add global config and verbose logs add InputState data struct for input handling and Input static class for user access

+++ engine/core/engine/layout.hpp
@@ -25,7 +25,7 @@ namespace s2f
     struct Layout
     {
         std::vector<LayoutElement> elements;
-        size_t stride{ 0 };
+        u64 stride{ 0 };
 	};
 
 	i32 getLayoutElementComponentCount(LayoutElementType type);
